Adding Profile requirenments
* Removed seeker context setting from marketing pages
* Removed sessions skipping
