+++ src/app/(app)/ContextProvider.jsx
@@ -1,14 +1,20 @@
 'use client';
-import { createContext, useState } from 'react';
+import { createContext, useEffect, useState } from 'react';
 
 export const ThemeContext = createContext('light');
 export const SeekerContext = createContext(null);
 export const EntryContext = createContext([]);
 
-export function ContextProvider({ children }) {
+export function ContextProvider({ children, getCurrentSession }) {
   const [theme, setTheme] = useState('light');
   const [seeker, setSeeker] = useState(null);
   const [entries, setEntries] = useState([]);
+
+  useEffect(() => {
+    getCurrentSession().then((data) => {
+      setSeeker(data.seeker);
+    });
+  }, [setSeeker]);
 
   return (
     <ThemeContext.Provider value={theme}>
