+++ src/app/(app)/layout.module.scss
@@ -42,7 +42,7 @@
       border-bottom: 1px solid var(--borderColor);
       padding: 0 20px;
 
-      search {
+      .search {
         background-color: var(--backgroundAccent);
         border: 2px solid var(--backgroundAccent);
         border-radius: 20px;
