+++ src/app/(landing)/(auth)/authLayout.module.scss
@@ -5,6 +5,7 @@
   width: 100vw;
   max-width: 450px;
   margin: 0 auto;
+  padding-top: 50px;
 
   & > h1 {
     font-size: 2rem;
@@ -28,6 +29,10 @@
       color: var(--accent);
       text-decoration: underline;
     }
+  }
+
+  @media (max-width: 500px) {
+    padding-top: 30px;
   }
 }
