--- src/app/(landing)/LandingLinks.jsx
@@ -1,32 +1,0 @@
-'use client';
-import { useContext, useEffect, useState } from 'react';
-import Link from 'next/link';
-import { SeekerContext } from '@/lib/helpers/ContextProvider.jsx';
-import styles from './marketingLayout.module.scss';
-
-export default function LandingLinks({ getCurrentSession }) {
-  const { seeker, setSeeker } = useContext(SeekerContext);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    if (seeker) return setIsLoading(false);
-    getCurrentSession().then((data) => {
-      setIsLoading(false);
-      setSeeker(data.seeker);
-    });
-  }, [seeker, setSeeker, getCurrentSession]);
-
-  return (
-    <>
-      {!isLoading && !seeker && (
-        <Link href='/start' className={styles.start}>
-          start<span> journey</span>
-        </Link>
-      )}
-
-      <Link href={seeker ? '/home' : '/continue'} className={styles.continue}>
-        continue
-      </Link>
-    </>
-  );
-}
