+++ src/app/layout.jsx
@@ -1,5 +1,4 @@
 import { Inter } from 'next/font/google';
-import { ContextProvider } from '@/lib/helpers/ContextProvider.jsx';
 import './globals.css';
 
 const inter = Inter({ subsets: ['latin'] });
@@ -7,7 +6,7 @@ const inter = Inter({ subsets: ['latin'] });
 export const metadata = {
   applicationName: 'Helios',
   title: 'Helios',
-  description: 'An assistive journaling app',
+  description: 'The journal that grows with you',
   keywords: ['journal', 'diary', 'write'],
   creator: 'Estefan Hu',
 };
@@ -15,9 +14,7 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang='en'>
-      <body className={inter.className}>
-        <ContextProvider>{children}</ContextProvider>
-      </body>
+      <body className={inter.className}>{children}</body>
     </html>
   );
 }
